# src/efa-server/infra/rest/generated/server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the spec that defines the API provided by the application\
    \ to register devices to a fabric, configure fabric parameters, validate all the\
    \ devices in the fabric and configure switches for IP Fabric with/without overlay"
  version: "1.0"
  title: "Simplified IP Fabric"
  contact:
    name: "Extreme Networks"
    url: "http://www.extremenetworks.com"
    email: "support@extremenetworks.com"
  license:
    name: "Apache license"
    url: "http://www.extremenetworks.com"
host: "localhost:8081"
basePath: "/v1"
schemes:
- "http"
paths:
...
  /fabric:
    get:
      tags:
      - "Fabric"
      summary: "getFabric"
      description: "Get only specified fabric details. The fabric can be identified\
        \ by id or name"
      operationId: "GetFabric"
      parameters:
      - name: "name"
        in: "query"
        description: "Name of the fabric to retrieve"
        required: true
        type: "string"
        x-exportParamName: "Name"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/FabricdataResponse"
        401:
          description: "Authorization information is missing or invalid."
        404:
          description: "A fabric with the specified name was not found."
        500:
          description: "Unexpected error."
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/ErrorModel"
    post:
      tags:
      - "Fabric"
      summary: "Create a Fabric"
      operationId: "createFabric"
      parameters:
      - in: "body"
        name: "fabric"
        description: "Add a new Fabric."
        required: false
        schema:
          $ref: "#/definitions/NewFabric"
        x-exportParamName: "Fabric"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/FabricdataResponse"
        401:
          description: "Authorization information is missing or invalid."
        409:
          description: "A fabric with the specified name already exists."
        500:
          description: "Unexpected error."
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/ErrorModel"
    put:
      tags:
      - "Fabric"
      summary: "Update a Fabric settings"
      operationId: "updateFabric"
      parameters:
      - in: "body"
        name: "fabric_settings"
        description: "Update Fabric Settings."
        required: false
        schema:
          $ref: "#/definitions/FabricSettings"
        x-exportParamName: "FabricSettings"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/FabricdataResponse"
        400:
          description: "Incorrect values specified for Fabric setting"
        401:
          description: "Authorization information is missing or invalid."
        404:
          description: "A fabric with the specified name was not found."
        409:
          description: "A fabric settings already exist and cannot be updated."
        500:
          description: "Unexpected error."
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/FabricdataErrorResponse"
    delete:
      tags:
      - "Fabric"
      summary: "deleteFabric"
      description: "Delete the specified fabric."
      operationId: "DeleteFabric"
      parameters:
      - name: "name"
        in: "query"
        description: "Name of the fabric to be deleted"
        required: true
        type: "string"
        x-exportParamName: "Name"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/FabricdataResponse"
        401:
          description: "Authorization information is missing or invalid."
        404:
          description: "A fabric with the specified name was not found."
        500:
          description: "Unexpected error."
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/ErrorModel"
...
definitions:
  NewFabric:
    required:
    - "name"
    properties:
      name:
        type: "string"
    example:
      name: "name"
  FabricParameter:
    required:
    - "key"
    - "value"
    properties:
      key:
        type: "string"
      value:
        type: "string"
    example:
      value: "value"
      key: "key"
  FabricSettings:
    required:
    - "keyval"
    - "name"
    properties:
      name:
        type: "string"
      keyval:
        type: "array"
        items:
          $ref: "#/definitions/FabricParameter"
    example:
      name: "name"
      keyval:
      - value: "value"
        key: "key"
      - value: "value"
        key: "key"
...
  FabricdataResponse:
    type: "object"
    properties:
      fabric_name:
        type: "string"
        example: "default"
        description: "Name of the fabric"
      fabric_id:
        type: "integer"
        format: "int32"
        example: 1
        description: "Database ID of the fabric"
      fabric_settings:
        type: "object"
        additionalProperties:
          type: "string"
    title: "fabricdata response"
    example:
      fabric_name: "default"
      fabric_id: 1
      fabric_settings:
        key: "fabric_settings"
...
  FabricdataErrorResponse:
    type: "object"
    properties:
      fabric_name:
        type: "string"
        example: "default"
        description: "Name of the fabric"
      fabric_id:
        type: "integer"
        format: "int32"
        example: 1
        description: "Database ID of the fabric"
      fabric_settings:
        type: "object"
        additionalProperties:
          type: "string"
    title: "fabric errordata response"
...
  ErrorModel:
    type: "object"
    required:
    - "code"
    - "message"
    properties:
      message:
        type: "string"
      code:
        type: "integer"
        minimum: 100
        maximum: 600
...