# src/efa-server/infra/rest/generated/server/go/fabric_settings.go
/*
 * Simplified IP Fabric
 *
 * This is the spec that defines the API provided by the application to register devices to a fabric, configure fabric parameters, validate all the devices in the fabric and configure switches for IP Fabric with/without overlay
 *
 * API version: 1.0
 * Contact: support@extremenetworks.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type FabricSettings struct {

	Name string `json:"name"`

	Keyval []FabricParameter `json:"keyval"`
}